+++ oci/database_maintenance_runs_data_source.go
@@ -157,9 +157,15 @@ func (s *DatabaseMaintenanceRunsDataSourceCrud) SetData() error {
 
 		maintenanceRun["maintenance_type"] = r.MaintenanceType
 
+		if r.PatchFailureCount != nil {
+			maintenanceRun["patch_failure_count"] = *r.PatchFailureCount
+		}
+
 		if r.PatchId != nil {
 			maintenanceRun["patch_id"] = *r.PatchId
 		}
+
+		maintenanceRun["patching_mode"] = r.PatchingMode
 
 		if r.PeerMaintenanceRunId != nil {
 			maintenanceRun["peer_maintenance_run_id"] = *r.PeerMaintenanceRunId